update createPrj.py add VSCode default config add some diy config

+++ apps/createPrj.py
@@ -3,12 +3,15 @@ import os
 import shutil
 from datetime import datetime
 
+mInit = True
+
 def createDir(rootDir,dirList):
     for dir in dirList:
         folder_path = os.path.join(rootDir,dir)
         os.makedirs(folder_path)
-        init_file_path = os.path.join(folder_path, "init")
-        open(init_file_path, "a").close()
+        if mInit:
+            init_file_path = os.path.join(folder_path, "init")
+            open(init_file_path, "a").close()
         
 def writeFile(file_path,content):
     with open(file_path, "w", encoding="utf-8") as file:
@@ -17,29 +20,49 @@ def writeFile(file_path,content):
 def touch_file(file_path):
     with open(file_path, 'a'):
         os.utime(file_path, None)
+        
+print("\nAuthor: Ricken - ricken.cn")
+print("Update: 2023/12/23 21:12")
+print("About:  A script for quickly building the Cdroid project framework\n")
 
-# 获取用户输入的文件夹名称
-project_name = input("请输入项目名称(英文且勿使用空格)：")
+project_name = ""
+while True:
+    # 获取用户输入的文件夹名称
+    project_name = input("请输入项目名称(英文且勿使用空格以及-)：")
+    
+    if ' ' in project_name or '-' in project_name or not project_name:
+        print("项目名称格式非法")
+        continue
 
-# 判断文件夹是否存在
-if os.path.exists(project_name):
-    # 提示用户文件夹已存在，并询问是否删除
-    choice = input(f"项目 '{project_name}' 已存在，是否删除并重新创建？(y/n): ")
+    # 判断文件夹是否存在
+    if os.path.exists(project_name):
+        # 提示用户文件夹已存在，并询问是否删除
+        choice = input(f"项目 '{project_name}' 已存在!!!\nY:删除并重新创建 R:重新输入 N:退出 : ")
 
-    if choice.lower() == "y":
-        # 删除已存在的文件夹
-        shutil.rmtree(project_name)
+        if choice.lower() == "y":
+            # 删除已存在的文件夹
+            shutil.rmtree(project_name)
+            break
+        elif choice.lower() == "r":
+            print("")
+            continue
+        else:
+            print("取消项目创建。")
+            exit()
     else:
-        print("取消项目创建。")
-        exit()
+        break
+
+mInitType = input(f"是否为每个目录添加空白文件进行占位(Y:是|默认 N:否): ")
+if mInitType.lower() == "n":
+    mInit = False
 
 # 创建文件夹
 os.makedirs(project_name)
-typeDir = ['3rdpart','assets','common','conf','ctrl','docs','net','src']
+typeDir   = ['.vscode','3rdpart','assets','conf','docs','fonts','src']
 createDir(project_name,typeDir)
 assetsDir = ['anim','color','drawable','layout','mipmap','values']
 createDir(os.path.join(project_name,'assets'),assetsDir)
-srcDir = ['data','function','viewlibs','page']
+srcDir    = ['common','data','function','net','page','protocol','uart','viewlibs']
 createDir(os.path.join(project_name,'src'),srcDir)
 
 img = "useless"
@@ -73,16 +96,16 @@ HARD_VER=A0.$HARD_VER_MAJ_NO.$HARD_VER_MIN_NO
 DATE=$(date +"%Y-%m-%d %H:%M.%S %Z")
 VER_DATE=$(date +%g).$(date +%v).$(date +%j)
 
-VER="CDROID_V${v1}.${v2}.${v3}.${v4}"
+VER="${v1}.${v2}.${v3}.${v4}"
 
-ln -sf $PWD/conf/cdroid_series_conf.h  common/hv_series_conf.h
-ln -sf $PWD/conf/defualt_config.h  common/hv_defualt_config.h
+ln -sf $PWD/conf/cdroid_series_conf.h  src/common/hv_series_conf.h
+ln -sf $PWD/conf/defualt_config.h  src/common/hv_defualt_config.h
 
 echo "set version $VER"
 echo "set telnet version  $VER_DATE"
 echo "set version TIME $DATE"
 version_info=`cat conf/version.h | sed "s/RevTag/$VER/g" | sed "s/RevTIME/$DATE/g" | sed "s/HV_MAJ/$SOFT_VER_MAJ_NO/g" | sed "s/HV_MIN/$SOFT_VER_MIN_NO/g" | sed "s/RevGITLAB/$GITLAB/g" | sed "s/RevHardVer/$HARD_VER/g"`
-echo "$version_info" > common/hv_version.h
+echo "$version_info" > src/common/hv_version.h
 if test -d packfs ; then
 version_system=`cat packfs/hv_system_conf | sed -r "s/.*HV_TELNET_VER.+/HV_TELNET_VER=$VER_DATE/g" | sed -r "s/.*HV_VERSION_NUM.+/HV_VERSION_NUM=$VER_MAIN_NO/g" | sed -r "s/.*HV_VERSION_CODE.+/HV_VERSION_CODE=V$VER_MAJ_NO.$VER_MIN_NO/g" | sed -r "s:.*HV_OUTPUT_PATH.+:HV_OUTPUT_PATH=$WORK_APP_OUT_PATH:g"`
 echo "$version_system" > packfs/hv_system_conf
@@ -103,14 +126,15 @@ set(CMAKE_CXX_STANDARD 11)
 include_directories(
     ./
     ./3rdparty
-    ./common
-    ./controller
-    ./net
 
+    ./src/common
     ./src/data
     ./src/function
-    ./src/viewlibs
+    ./src/net
     ./src/page
+    ./src/protocol
+    ./src/uart
+    ./src/viewlibs
 
     ${CDROID_INCLUDE_DIRS}
     ${CDROID_DEPINCLUDES}
@@ -129,11 +153,15 @@ elseif(CDROID_CHIPSET STREQUAL x64)
     add_definitions(-DCDROID_X64)
 endif()
 
-aux_source_directory(./net COLMO_SRC_UTIL)
-aux_source_directory(./common COLMO_SRC_COMM)
-aux_source_directory(./controller COLMO_SRC_CTRL)
-aux_source_directory(./src/page COLMO_SRC_CTRL)
-file(GLOB_RECURSE COLMO_SRCS ${CMAKE_CURRENT_SOURCE_DIR} *.cc)
+aux_source_directory(./src/common PRJ_SRC_COMM)
+aux_source_directory(./src/data PRJ_SRC_DATA)
+aux_source_directory(./src/function PRJ_SRC_FUN)
+aux_source_directory(./src/net PRJ_SRC_NET)
+aux_source_directory(./src/page PRJ_SRC_PAGE)
+aux_source_directory(./src/protocol PRJ_SRC_PROTOCOL)
+aux_source_directory(./src/uart PRJ_SRC_UART)
+aux_source_directory(./src/viewlibs PRJ_SRC_VIEWLIB)
+file(GLOB_RECURSE PRJ_SRCS ${CMAKE_CURRENT_SOURCE_DIR} *.cc)
 
 
 # add_definitions(--include cdtypes.h)
@@ -141,7 +169,7 @@ file(GLOB_RECURSE COLMO_SRCS ${CMAKE_CURRENT_SOURCE_DIR} *.cc)
 
 message("PROJECT_NAME=${PROJECT_NAME}")
 
-add_executable(${PROJECT_NAME} ${COLMO_SRCS} ${COLMO_SRC_COMM} ${COLMO_SRC_UTIL} ${COLMO_SRC_CTRL})
+add_executable(${PROJECT_NAME} ${PRJ_SRCS} ${PRJ_SRC_COMM} ${PRJ_SRC_DATA} ${PRJ_SRC_FUN} ${PRJ_SRC_NET} ${PRJ_SRC_PAGE} ${PRJ_SRC_PROTOCOL} ${PRJ_SRC_UART} ${PRJ_SRC_VIEWLIB})
 
 
 CreatePAK(${PROJECT_NAME} ${PROJECT_SOURCE_DIR}/assets ${PROJECT_BINARY_DIR}/${PROJECT_NAME}.pak ${PROJECT_SOURCE_DIR}/R.h)
@@ -168,21 +196,22 @@ __CREATETIME__
 - - anim      动画
 - - color     颜色
 - - drawable  样式
-- - fonts     字体
 - - layout    布局
 - - mipmap    图片
 - - strings   文本
 - - values    数据
-- common   一些常用函数
-- conf     配置文件
-- ctrl     通信
-- docs     说明文档
-- net      网络
+- conf     配置文件(touch后编译生效,文件生成于src/common)
+- docs     项目文档
+- fonts    项目字体
 - src      功能代码
+- - common    常用函数
 - - data      数据存储
 - - function  功能函数
+- - net       网络相关
+- - page      页面函数
+- - protocol  通信协议
+- - uart      串口相关
 - - viewlibs  自定义页面
-- - page      页面
 ~~~
 
 ## 声明
@@ -289,24 +318,28 @@ mVersionH = '''#ifndef __CONF_VERSION_H__
 #define HV_HARD_VERSION "V" HV_CPU_NAME "." HV_HARD_VER_S
 
 
+/// @brief 获取格式化后的软件版本，用于展示
+/// @param _szBuf 
+/// @return 
 static inline char* HV_FORMAT_VERSION_STRING(char* _szBuf)
 {
 	sprintf(_szBuf, "%s_V%s",HV_SERIES_NAME,HV_SOFT_VER_EXT);
 	return _szBuf;
 }
 
-
+/// @brief 获取当前版本构建时间
+/// @return 
 static inline const char* HV_FORMAT_VER_TIME_STRING()
 {
 	return HV_SOFT_VER_TIME;
 }
 
-
+/// @brief 获取当前版本号
+/// @return 
 static inline const char* HV_FORMAT_VER_EXT_STRING()
 {
 	return HV_SOFT_VER_EXT;
 }
-
 
 static inline const char* HV_FORMAT_GIT_HARD_STRING()
 {
@@ -334,8 +367,7 @@ writeFile(os.path.join(project_name,'conf','defualt_config.h'),mDefaultConfig);
 mCdroidSeries = '''#ifndef __HV_SERIES_CONFIG_H__
 #define __HV_SERIES_CONFIG_H__
 
-#define HV_SERIES_NAME	"FUGE_CDROID"
-#define HV_MODEL_FUGE_CDROID0	1
+#define HV_SERIES_NAME	"CDROID"
 
 #define HV_FLASH_SIZE	"128NR"
 #define HV_FLASH_16NR	1
@@ -362,6 +394,39 @@ mCdroidSeries = '''#ifndef __HV_SERIES_CONFIG_H__
 '''
 
 writeFile(os.path.join(project_name,'conf','cdroid_series_conf.h'),mCdroidSeries);
+
+
+mVSCodeConfig = """{
+    "configurations": [
+        {
+            "name": "Linux",
+            "includePath": [
+                "${workspaceFolder}/*",
+                "${workspaceFolder}/**",
+                "${workspaceFolder}/src/common/**",
+                "${workspaceFolder}/src/data/**",
+                "${workspaceFolder}/src/function/**",
+                "${workspaceFolder}/src/net/**",
+                "${workspaceFolder}/src/page/**",
+                "${workspaceFolder}/src/protocol/**",
+                "${workspaceFolder}/src/uart/**",
+                "${workspaceFolder}/src/viewlibs/**",
+                "~/cdroid/outX64-Debug/include/**",
+                "~/cdroid/outX64-Debug/include/gui/**",
+                "~/cdroid/src/**",
+                "/opt/vcpkg/packages/**"
+            ],
+            "defines": [],
+            "compilerPath": "/usr/bin/gcc",
+            "cStandard": "c17",
+            "cppStandard": "gnu++17",
+            "intelliSenseMode": "linux-gcc-x64"
+        }
+    ],
+    "version": 4
+}"""
+
+writeFile(os.path.join(project_name,'.vscode','c_cpp_properties.json'),mVSCodeConfig);
 
 
 touch_file("CMakeLists.txt")
